+++ champi_interfaces/CMakeLists.txt
@@ -10,11 +10,12 @@ find_package(ament_cmake REQUIRED)
 find_package(geometry_msgs REQUIRED)
 find_package(rosidl_default_generators REQUIRED)
 
-
-
 rosidl_generate_interfaces(${PROJECT_NAME}
-        "action/Navigate.action"
-        DEPENDENCIES geometry_msgs
+  "msg/ChampiPoint.msg"
+  "msg/ChampiSegment.msg"
+  "msg/ChampiPath.msg"
+  "action/Navigate.action"
+  DEPENDENCIES geometry_msgs
 )
 
 if(BUILD_TESTING)
