+++ champi_interfaces/msg/ChampiPath.msg
@@ -1,0 +1,3 @@
+string name
+ChampiSegment[] segments
+float32 max_time_allowed
