+++ champi_interfaces/msg/ChampiPoint.msg
@@ -1,0 +1,5 @@
+string name
+geometry_msgs/Pose pose
+int32 point_type
+float32 tolerance
+bool robot_should_stop_here
