+++ champi_interfaces/msg/ChampiSegment.msg
@@ -1,0 +1,5 @@
+string name
+ChampiPoint start
+ChampiPoint end
+float32 speed
+ChampiPoint look_at_point
