+++ champi_interfaces/package.xml
@@ -19,6 +19,8 @@
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
 
+  <exec_depend>rosidl_default_runtime</exec_depend>
+
   <export>
     <build_type>ament_cmake</build_type>
   </export>
